# src/build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"Record_string.string_": {
				"properties": {},
				"additionalProperties": {
					"type": "string"
				},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"LanguageMap": {
				"$ref": "#/components/schemas/Record_string.string_",
				"description": "RFC 5646 language-tagged string map. Keys are language tags, values are human-readable strings."
			},
...
			"InteractionType": {
				"type": "string",
				"enum": [
					"true-false",
					"choice",
					"fill-in",
					"long-fill-in",
					"matching",
					"performance",
					"sequencing",
					"likert",
					"numeric",
					"other"
				]
			},
			"InteractionComponent": {
				"properties": {
					"id": {
						"type": "string"
					},
					"description": {
						"$ref": "#/components/schemas/LanguageMap"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Record_string.unknown_": {
				"properties": {},
				"additionalProperties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"Extensions": {
				"$ref": "#/components/schemas/Record_string.unknown_",
				"description": "Open extension map. Keys are IRIs, values are arbitrary JSON."
			},
			"ActivityDefinition": {
				"properties": {
					"name": {
						"$ref": "#/components/schemas/LanguageMap"
					},
					"description": {
						"$ref": "#/components/schemas/LanguageMap"
					},
					"type": {
						"type": "string"
					},
					"moreInfo": {
						"type": "string"
					},
					"interactionType": {
						"$ref": "#/components/schemas/InteractionType"
					},
					"correctResponsesPattern": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"choices": {
						"items": {
							"$ref": "#/components/schemas/InteractionComponent"
						},
						"type": "array"
					},
					"scale": {
						"items": {
							"$ref": "#/components/schemas/InteractionComponent"
						},
						"type": "array"
					},
					"source": {
						"items": {
							"$ref": "#/components/schemas/InteractionComponent"
						},
						"type": "array"
					},
					"target": {
						"items": {
							"$ref": "#/components/schemas/InteractionComponent"
						},
						"type": "array"
					},
					"steps": {
						"items": {
							"$ref": "#/components/schemas/InteractionComponent"
						},
						"type": "array"
					},
					"extensions": {
						"$ref": "#/components/schemas/Extensions"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Activity": {
				"properties": {
					"objectType": {
						"type": "string",
						"enum": [
							"Activity"
						],
						"nullable": false
					},
					"id": {
						"type": "string"
					},
					"definition": {
						"$ref": "#/components/schemas/ActivityDefinition"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			},
			"xapi_basic": {
				"type": "http",
				"scheme": "basic"
			}
		}
	},
	"info": {
		"title": "xAPI LRS",
		"version": "2.0.0",
		"contact": {}
	},
	"paths": {
...
		"/xapi/activities/state": {
			"get": {
				"operationId": "GetState",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"type": "string"
											},
											"type": "array"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"xAPI State"
				],
				"security": [
					{
						"jwt": []
					},
					{
						"xapi_basic": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "activityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "agent",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "stateId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "registration",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "since",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "PutState",
				"responses": {
					"204": {
						"description": "No Content"
					}
				},
				"tags": [
					"xAPI State"
				],
				"security": [
					{
						"jwt": []
					},
					{
						"xapi_basic": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "activityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "agent",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "stateId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "registration",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "If-Match",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "If-None-Match",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "PostState",
				"responses": {
					"204": {
						"description": "No Content"
					}
				},
				"tags": [
					"xAPI State"
				],
				"security": [
					{
						"jwt": []
					},
					{
						"xapi_basic": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "activityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "agent",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "stateId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "registration",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "If-Match",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "If-None-Match",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteState",
				"responses": {
					"204": {
						"description": "No Content"
					}
				},
				"tags": [
					"xAPI State"
				],
				"security": [
					{
						"jwt": []
					},
					{
						"xapi_basic": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "activityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "agent",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "stateId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "registration",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "If-Match",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/xapi/activities": {
			"get": {
				"operationId": "GetActivity",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Activity"
								}
							}
						}
					}
				},
				"description": "GET /xapi/activities\n\nReturns the canonical Activity object for the given activityId.\nPer spec, if the LRS has no canonical definition it returns\nan Activity with just the id.",
				"tags": [
					"xAPI Activities"
				],
				"security": [
					{
						"jwt": []
					},
					{
						"xapi_basic": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "activityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}